L08: work with GC

+++ L08-gc/homework/src/main/java/ru/calculator/CalcDemo.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 
 public class CalcDemo {
     public static void main(String[] args) {
-        long counter = 100_000_000;
+        int counter = 100_000_000;
         var summator = new Summator();
         long startTime = System.currentTimeMillis();
